Make role permissions form dynamic
- display the help text in a tooltip
- disable the inherited roles with a `inherited` label
- only display the tournament selector if a tournament scope role is selected
