+++ src/web/settings.py
@@ -26,7 +26,6 @@ from common.i18n import gettext, ngettext
 from data.input_output import OnlineDataSourceManager
 
 from plugins.manager import plugin_manager
-from web.controllers.admin.device_admin_controller import DeviceAdminController
 from web.controllers.admin.display_controller_admin_controller import (
     DisplayControllerAdminController,
 )
@@ -41,7 +40,7 @@ from web.controllers.admin.screen_admin_controller import ScreenAdminController
 from web.controllers.admin.pairings_admin_controller import PairingsAdminController
 from web.controllers.admin.timer_admin_controller import TimerAdminController
 from web.controllers.admin.tournament_admin_controller import TournamentAdminController
-from web.controllers.admin.account_admin_controller import AccountAdminController
+from web.controllers.admin.access_admin_controller import AccessAdminController
 from web.controllers.background_controller import BackgroundController
 from web.controllers.index_controller import IndexController
 from web.controllers.qrcode_controller import QRCodeController
@@ -86,8 +85,7 @@ route_handlers: Sequence[ControllerRouterHandler] = [
     PlayerAdminController,
     DisplayControllerAdminController,
     QRCodeController,
-    AccountAdminController,
-    DeviceAdminController,
+    AccessAdminController,
     ProfileController,
     static_files_router,
     # Plugin controllers
