+++ src/web/templates/common/base.html
@@ -407,6 +407,12 @@
                 // Allow the tooltip content to have a style attribute
                 bootstrap.Tooltip.Default.allowList['*'].push('style')
             }
+
+            $(document).on('select2:close', '.sharly-chess-select2', function (e) {
+                var evt = "scroll.select2";
+                $(e.target).parents().off(evt);
+                $(window).off(evt);
+            })
         </script>
     {% endif %}
 </html>
